Removal of the old gitlab file and added a basic github workflow yml file. This is based loosely off the sample python package one.
Extended support to also test on Windows and MacOS as well as test against Python 3.9. Updated the pipeline badge in the README

Added in section to upload the robot log file. Once confirmed it is working, a line will be added to only upload this on failure.

Whitespace syntax fix

Updated tox file to specify the output file paths and removed a rogue file

Attempted fix for a name collision when uploading robot log files as artifacts

Included the os type in the log file name to avoid a conflict

Renamed the job name and added a condition to only upload robot log files on failure

Fixed a value in the GH workflow that was at the wrong level and added a hyperlink around the build pipeline badge