Migrated CI pipeline from GitLab to GitHub Actions to allow forked repos to run test pipelines (#3)
* Removal of the old gitlab file and added a basic github workflow yml file. This is based loosely off the sample python package one.

Extended support to also test on Windows and MacOS as well as test against Python 3.9. Updated the pipeline badge in the README

Added in section to upload the robot log file. Once confirmed it is working, a line will be added to only upload this on failure.

Whitespace syntax fix

Updated tox file to specify the output file paths and removed a rogue file

Attempted fix for a name collision when uploading robot log files as artifacts

Included the os type in the log file name to avoid a conflict

Renamed the job name and added a condition to only upload robot log files on failure

Fixed a value in the GH workflow that was at the wrong level and added a hyperlink around the build pipeline badge

* Added a newline at the bottom of the workflow yml file (as per GitHub's request)

* PR Comments - Shortened the name of the test workflow and added an empty newline to the bottom of the tox.ini file

* Set pinned versions for the dev dependencies and removed the explicit install from the tox.ini which isn't required.

Co-authored-by: Chris Brookes <dev@net.com>